requirements.txt
flask
flask-cors
python-dotenv
pandas
numpy
scikit-learn
imbalanced-learn
joblib
matplotlib
seaborn
jupyter
psycopg2-binary

